Draw the photographed rainbows on the globe.
To accomplish this, we need a bridge between the app and the webview.
First we convert the info from the photos into the ‘GEOJSON’ format .
Than we pass these coordinates and the commando to draw this into the webview context:

globe.evalJS('svg.append("path").datum(' + JSON.stringify(features2Photos())  + ').attr("d", path).attr("class", "place");');

Almost surprised that this works.