Somewhat better prepared for multiple screen sizes
We started designing everything for the iPhone 4s—
We had already made some views work for the iPad,
but now the photo scroll works as well.

The settings screen is no primarily designed for larger phones:
on the 4s you have to scroll. Some extra space added to
accommodate for the toolbar.