Halfway decent photo-scroll
No longer need the time-out, as the photoUpdate no longer is triggered on tab switch

The scrollToIndex did not work together with the tabIndex property: it works on the
actual order of rows.

As the row starts with a separator row, then photo rows and separator rows intermittently,
We know index n in the photos corresponds with index 2 * n + 1 in the table.

FIXME: it somehow not always scrolls to the top of the photo (i.e. in the Moscow case)